requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.3
black==21.7b0
bs4==0.0.1
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==8.0.1
firebase-admin==5.0.1
google-api-core==1.31.1
google-api-python-client==2.15.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
google-cloud-core==1.7.2
google-cloud-firestore==2.2.0
google-cloud-storage==1.41.1
google-crc32c==1.1.2
google-resumable-media==1.3.2
googleapis-common-protos==1.53.0
grpcio==1.39.0
httplib2==0.19.1
idna==3.2
importlib-metadata==4.6.1
msgpack==1.0.2
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.9.0
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
pytz==2021.1
regex==2021.7.6
requests==2.26.0
rsa==4.7.2
six==1.16.0
soupsieve==2.2.1
tomli==1.2.0
typed-ast==1.4.3
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
zipp==3.5.0
